Fixed a 100AP conversion issue for morale and suppression fire. Fixed a CTD caused by trying to "pickup" an item from sector inventory while simultaneously looking at that items description box. Upgraded an INT8 to INT16 in training functions to try and catch a problem with INI values being set too high.
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@2419 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Strategic/Assignments.cpp
@@ -3731,7 +3731,7 @@ INT16 GetBonusTrainingPtsDueToInstructor( SOLDIERTYPE *pInstructor, SOLDIERTYPE 
 	INT8 bTraineeSkill = 0;
 	INT8 bTrainerEffSkill = 0;
 	INT8 bTrainerNatSkill = 0;
-	INT8 bTrainingBonus = 0;
+	INT16 bTrainingBonus = 0;
 	INT8 bOpinionFactor;
 
 	// assume training impossible for max pts
@@ -3916,7 +3916,7 @@ INT16 GetBonusTrainingPtsDueToInstructor( SOLDIERTYPE *pInstructor, SOLDIERTYPE 
 INT16 GetSoldierTrainingPts( SOLDIERTYPE *pSoldier, INT8 bTrainStat, BOOLEAN fAtGunRange, UINT16 *pusMaxPts )
 {
 	INT16 sTrainingPts = 0;
-	INT8	bTrainingBonus = 0;
+	INT16	bTrainingBonus = 0;
 	INT8	bSkill = 0;
 
 	// assume training impossible for max pts
@@ -3992,7 +3992,7 @@ INT16 GetSoldierTrainingPts( SOLDIERTYPE *pSoldier, INT8 bTrainStat, BOOLEAN fAt
 INT16 GetSoldierStudentPts( SOLDIERTYPE *pSoldier, INT8 bTrainStat, BOOLEAN fAtGunRange, UINT16 *pusMaxPts )
 {
 	INT16 sTrainingPts = 0;
-	INT8	bTrainingBonus = 0;
+	INT16	bTrainingBonus = 0;
 	INT8	bSkill = 0;
 
 	INT16 sBestTrainingPts, sTrainingPtsDueToInstructor;
@@ -4228,7 +4228,7 @@ BOOLEAN TrainTownInSector( SOLDIERTYPE *pTrainer, INT16 sMapX, INT16 sMapY, INT1
 INT16 GetTownTrainPtsForCharacter( SOLDIERTYPE *pTrainer, UINT16 *pusMaxPts )
 {
 	INT16 sTotalTrainingPts = 0;
-	INT8 bTrainingBonus = 0;
+	INT16 bTrainingBonus = 0;
 //	UINT8 ubTownId = 0;
 
 	// calculate normal training pts - what it would be if his stats were "normal" (ignoring drugs, fatigue)
